+++ info.c
@@ -3,7 +3,7 @@
 #include <unistd.h>
 #include <time.h>
 
-void checkit(float income, float expense, double  s){
+void checkit(double income, double expense, double  s){
 
 	long the_time;
 	time(&the_time);
@@ -11,7 +11,7 @@ void checkit(float income, float expense, double  s){
 	printf("%lf\n",s);
 	
 
-	fprintf(fp,"The user %s of home directory %s recorded an income of %f and expense of %f on %s Net balance: %lf\n",getlogin(),getenv("HOME"),income, expense,ctime(&the_time),s);
+	fprintf(fp,"The user %s of home directory %s recorded an income of %lf and expense of %lf on %s Net balance: %lf\n",getlogin(),getenv("HOME"),income, expense,ctime(&the_time),s);
 
 	fclose(fp);
 }
